Golang WebScrapper/llm/requirements.txt
replicate==1.0.4
python-dotenv==1.0.0
requests==2.31.0